+++ src/main/web_src/src/components/bestillingsveileder/stegVelger/StegVelger.js
@@ -30,7 +30,6 @@ export const StegVelger = ({ initialValues, onSubmit, children }) => {
 
 	const _handleSubmit = (values, formikBag) => {
 		const { setSubmitting } = formikBag
-		console.log('values :>> ', values)
 		if (!isLastStep()) {
 			setSubmitting(false)
 			handleNext()
@@ -59,7 +58,7 @@ export const StegVelger = ({ initialValues, onSubmit, children }) => {
 
 						<CurrentStepComponent formikBag={formikBag} stateModifier={stateModifier} />
 
-						<DisplayFormikState {...formikBag} />
+						{/* <DisplayFormikState {...formikBag} /> */}
 
 						<Navigation
 							showPrevious={step > 0}
